+++ odoo/local-src/specific_zetes/tools/domain_itemmove.py
@@ -243,6 +243,7 @@ class Itemmove(DomainInterface):
                     pack_op.picking_id.validate_picking()
 
         except Exception as e:
+            self.rollback_to_savepoint()
             _logger.error(str(e))
             params.log(picking_id=pack_op.picking_id.id,
                        operation_id=pack_operation_id,